accessory resolvers and schema

+++ graphql/resolvers/product.js
@@ -1,11 +1,13 @@
 const Product = require("../../models/product");
+const RentProduct = require("../../models/rent-product");
+const ProductQuotation = require("../../models/product-quotation");
 
 const { transformProduct } = require("./merge");
 
 module.exports = {
   products: async () => {
     try {
-      const products = await Product.find();
+      const products = await Product.find({ deleted: false });
       return products.map(product => {
         return transformProduct(product);
       });
@@ -25,7 +27,7 @@ module.exports = {
 
   subcategoryProducts: async args => {
     try {
-      const products = await Product.find({subcategories: args.id});
+      const products = await Product.find({ subcategories: args.id });
       return products.map(product => {
         return transformProduct(product);
       });
@@ -65,8 +67,20 @@ module.exports = {
 
   deleteProduct: async args => {
     try {
-      const product = await Product.findByIdAndDelete(args.id);
-      return transformProduct(product);
+      const rentProducts = await RentProduct.find({ product: args.id });
+      const productQuotations = await ProductQuotation.find({ product: args.id })
+
+      if (rentProducts || productQuotations) {
+        const product = await Product.findById(args.id)
+        product.subcategories = null;
+        product.deleted = true;
+        const result = await product.save()
+        return transformProduct(result);
+      } else {
+        const product = await Product.findByIdAndDelete(args.id);
+        return transformProduct(product);
+      }
+
     } catch (err) {
       throw err;
     }
